# Fix: CMake optimization flags unintentionally overwritten