## Description of the changes

This PR fixes an issue in `CMakeLists.txt` where compiler optimization flags are unintentionally overwritten.

### Problem

There is an issue in `CMakeLists.txt` where `CMAKE_CXX_FLAGS_RELEASE` is overwritten instead of appended to, causing all optimization flags to be lost.

Lines 67-73 in CMakeLists.txt:
```cmake
set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS} -Ofast -ffast-math -DNDEBUG")
if(APPLE AND CMAKE_SYSTEM_PROCESSOR STREQUAL "arm64")
  # Apple Sillicon (m1) does not support `march=native`. use `mcpu` instead.
  set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS} -mcpu=apple-m1")
else()
  set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS} -march=native")
endif()
```

In the conditional statements, `CMAKE_CXX_FLAGS_RELEASE` gets overwritten, losing all the optimization flags (`-Ofast -ffast-math -DNDEBUG`) that were set earlier.

### Changes Made

1. **Fixed flag overwriting**: Changed `${CMAKE_CXX_FLAGS}` to `${CMAKE_CXX_FLAGS_RELEASE}` to properly append architecture-specific flags
2. **Removed `-Werror`**: Removed to address compilation warnings that appear with proper optimization flags enabled

## Performance Impact

This fix provides a ~4x performance improvement in YCSB benchmarks:

**Before fix (flags overwritten):**
```bash
noxy@XG6326-2U:~/LineairDB/build/bench$ ./ycsb --thread 30 --clients 30 --records 100000 --contention 0.9 --epoch 1000
[Thread 276997] [2025-08-03 22:52:29.596] [info] [database_impl.h:53] LineairDB instance has been constructed. [+0msec]
[Thread 276997] [2025-08-03 22:52:29.599] [info] [benchmark.cpp:89] YCSB: Database population queries are enqueued [+3msec]
[Thread 276997] [2025-08-03 22:52:31.596] [info] [benchmark.cpp:94] YCSB: Database population is completed [+1996msec]
[Thread 276997] [2025-08-03 22:52:31.603] [info] [benchmark.cpp:235] YCSB: Benchmark start. [+6msec]
[Thread 276997] [2025-08-03 22:52:33.605] [info] [benchmark.cpp:245] YCSB: Benchmark end. [+2001msec]
[Thread 276997] [2025-08-03 22:52:38.598] [info] [benchmark.cpp:247] YCSB: DB Fenced. [+4993msec]
[Thread 276997] [2025-08-03 22:52:38.598] [info] [benchmark.cpp:261] YCSB: Benchmark completed. elapsed time: 2001ms, commits: 1050355, aborts: 469475, tps: 524915 [+0msec]
This benchmark result is saved into ycsb_result.json
[Thread 276997] [2025-08-03 22:52:44.598] [debug] [database_impl.h:76] Epoch number and Durable epoch number are ended at 16, and 0, respectively. [+6000msec]
[Thread 276997] [2025-08-03 22:52:44.598] [info] [database_impl.h:80] LineairDB instance has been destructed. [+0msec]
```

**After fix (proper optimization flags):**
```bash
noxy@XG6326-2U:~/LineairDB/build/bench$ ./ycsb --thread 30 --clients 30 --records 100000 --contention 0.9 --epoch 1000
[Thread 281862] [2025-08-03 22:59:11.050] [info] [database_impl.h:53] LineairDB instance has been constructed. [+0msec]
[Thread 281862] [2025-08-03 22:59:11.054] [info] [benchmark.cpp:89] YCSB: Database population queries are enqueued [+3msec]
[Thread 281862] [2025-08-03 22:59:13.051] [info] [benchmark.cpp:94] YCSB: Database population is completed [+1996msec]
[Thread 281862] [2025-08-03 22:59:13.059] [info] [benchmark.cpp:235] YCSB: Benchmark start. [+7msec]
[Thread 281862] [2025-08-03 22:59:15.060] [info] [benchmark.cpp:245] YCSB: Benchmark end. [+2001msec]
[Thread 281862] [2025-08-03 22:59:18.052] [info] [benchmark.cpp:247] YCSB: DB Fenced. [+2992msec]
[Thread 281862] [2025-08-03 22:59:18.052] [info] [benchmark.cpp:261] YCSB: Benchmark completed. elapsed time: 2001ms, commits: 4032400, aborts: 353478, tps: 2015192 [+0msec]
This benchmark result is saved into ycsb_result.json
[Thread 281862] [2025-08-03 22:59:24.052] [info] [database_impl.h:80] LineairDB instance has been destructed. [+6000msec]
```

## Discussion

With optimization flags properly enabled, the compiler generates warnings about memory allocation mismatches in `data_buffer.hpp` (mixing `new[]`/`delete[]` with `realloc`). This PR disables escalation of these warnings to errors by removing `-Werror`.

```bash
In file included from /home/noxy/LineairDB/src/types/data_item.hpp:30,
                 from /home/noxy/LineairDB/src/recovery/logger_base.h:20,
                 from /home/noxy/LineairDB/src/recovery/logger.h:25,
                 from /home/noxy/LineairDB/src/recovery/logger.cpp:17:
In member function ‘void LineairDB::DataBuffer::Reset(const std::byte*, size_t)’,
    inlined from ‘void LineairDB::DataBuffer::Reset(const std::string&)’ at /home/noxy/LineairDB/src/types/data_buffer.hpp:63:10,
    inlined from ‘LineairDB::WriteSetType LineairDB::Recovery::Logger::GetRecoverySetFromLogs(LineairDB::EpochNumber)’ at /home/noxy/LineairDB/src/recovery/logger.cpp:195:51:
/home/noxy/LineairDB/src/types/data_buffer.hpp:43:16: error: ‘void operator delete [](void*)’ called on a pointer passed to mismatched reallocation function ‘void* realloc(void*, size_t)’ [-Werror=mismatched-dealloc]
   43 |       delete[] value;
      |                ^~~~~
/home/noxy/LineairDB/src/types/data_buffer.hpp:52:25: note: call to ‘void* realloc(void*, size_t)’
   52 |             std::realloc(reinterpret_cast<void*>(value), s));
      |             ~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/noxy/LineairDB/src/types/data_buffer.hpp:43:16: error: ‘void operator delete [](void*)’ called on a pointer passed to mismatched reallocation function ‘void* realloc(void*, size_t)’ [-Werror=mismatched-dealloc]
   43 |       delete[] value;
      |                ^~~~~
/home/noxy/LineairDB/src/types/data_buffer.hpp:52:25: note: call to ‘void* realloc(void*, size_t)’
   52 |             std::realloc(reinterpret_cast<void*>(value), s));
      |             ~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
cc1plus: all warnings being treated as errors
```

There are two potential approaches:

1. **Current approach**: Remove `-Werror` and accept the warnings as non-critical
2. **Alternative approach**: Keep `-Werror` and fix the memory allocation issue in `data_buffer.hpp` (and also somewhere else, maybe)