@nikezono 2025-08-04
@Noxy3301 Thank you for the contribution! Great work.
But we need `-Werror` in CMakeLists.txt to keep "no warnings"  codebase.

I'll update your branch!